chore(future_builder): refactor and change file hierarchy
