+++ lib/pages/home/home.dart
@@ -2,12 +2,9 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:remote_guide_firebase/pages/home_guide/home_client.dart';
-import 'package:remote_guide_firebase/pages/profile/profile.dart';
-import 'package:remote_guide_firebase/pages/home_guide/upcoming/upcoming.dart';
+import 'package:remote_guide_firebase/pages/home/home_client.dart';
 import 'package:remote_guide_firebase/services/auth.dart';
 import 'package:remote_guide_firebase/services/database.dart';
-import 'package:remote_guide_firebase/pages/home_guide/my_routes/my_routes.dart';
 
 import 'home_guide.dart';
 
@@ -18,21 +15,9 @@ class Home extends StatefulWidget {
   State<Home> createState() => _Home();
 }
 
-class CustomUser {
-  final name;
-  final email;
-  final guide;
-  
-  CustomUser(this.guide, this.email, this.name);
-}
-
 class _Home extends State<Home> {
   late Future<DocumentSnapshot<Object?>?> _data;
-  late List<Widget> _widgetOptions = <Widget>[
-    MyRoutes(),
-    UpcomingTours(),
-    Profile()
-  ];
+  final AuthService _auth = AuthService();
 
   @override
   void initState() {
@@ -41,18 +26,6 @@ class _Home extends State<Home> {
     CollectionReference userCollection = FirebaseFirestore.instance.collection('users');
     _data = userCollection.doc(auth.currentUser?.email).get();
   }
-
-  int _selectedIndex = 0;
-  static const TextStyle optionStyle =
-  TextStyle(fontSize: 30, fontWeight: FontWeight.bold);
-
-  void _onItemTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-  }
-
-  final AuthService _auth = AuthService();
 
   @override
   Widget build(BuildContext context) {
@@ -94,28 +67,6 @@ class _Home extends State<Home> {
               }
             }
         )
-        // Center(
-        //       child: _widgetOptions.elementAt(_selectedIndex),
-        //   ),
-        // bottomNavigationBar: BottomNavigationBar(
-        //   items: const <BottomNavigationBarItem>[
-        //     BottomNavigationBarItem(
-        //       icon: Icon(Icons.list),
-        //       label: 'My routes',
-        //     ),
-        //     BottomNavigationBarItem(
-        //       icon: Icon(Icons.schedule),
-        //       label: 'Upcoming',
-        //     ),
-        //     BottomNavigationBarItem(
-        //       icon: Icon(Icons.person),
-        //       label: 'Profile',
-        //     ),
-        //   ],
-        //   currentIndex: _selectedIndex,
-        //   selectedItemColor: Colors.amber[800],
-        //   onTap: _onItemTapped,
-        // ),
     )
     );
   }
