+++ lib/pages/home/profile/profile.dart
@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:remote_guide_firebase/pages/profile/profile_info.dart';
+import 'package:remote_guide_firebase/pages/home/profile/profile_info.dart';
 
 class Profile extends StatefulWidget {
   const Profile({Key? key}) : super(key: key);
