+++ lib/pages/home/upcoming/upcoming.dart
@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
-import 'package:remote_guide_firebase/pages/home_guide/upcoming/bookings_list.dart';
+import 'package:remote_guide_firebase/pages/home/upcoming/bookings_list.dart';
 
 import '../my_routes/routes_list.dart';
