+++ lib/pages/models/tour_class.dart
