+++ lib/pages/wrapper.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:remote_guide_firebase/pages/authenticate/authenticate.dart';
-import 'home_guide/home.dart';
+import 'home/home.dart';
 import 'models/myuser.dart';
 
 class Wrapper extends StatelessWidget {
